requirements.txt
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
torch==2.3.0
torchvision==0.18.0
streamlit==1.35.0
streamlit-webrtc==0.47.6
av==12.0.0
